+++ src/containers/navigation/components/Miner/components/PoolStatsTile/styles.ts
@@ -56,25 +56,6 @@ export const BalanceVal = styled(Typography)`
     letter-spacing: -0.4px;
 `;
 
-export const TimerDot = styled.div`
-    border-radius: 50%;
-    background-color: #33cd7e;
-    width: 7px;
-    height: 7px;
-`;
-export const Timer = styled.div`
-    width: 100%;
-    gap: 3px;
-    display: flex;
-    align-items: center;
-    color: ${({ theme }) => theme.palette.text.contrast};
-    background-color: ${({ theme }) => theme.palette.contrast};
-    padding: 2px 7px;
-    border-radius: 50px;
-    line-height: 1;
-    font-weight: 500;
-`;
-
 export const TriggerWrapper = styled(m.div)`
     width: 14px;
     height: 14px;
